refactor: change package 'dto' from controller to api
