+++ src/main/java/br/com/fiap/tech_challenge/adapters/driver/api/dto/CustomerRequestDTO.java
@@ -1,4 +1,4 @@
-package br.com.fiap.tech_challenge.adapters.driver.api.controller.dto;
+package br.com.fiap.tech_challenge.adapters.driver.api.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
